storm: rearrange fmap for 8M
Storm flash map is reshuffled to fully utilize the 8MB bootptom
installed on the device.

The lower half is read only and is not changing much, just using reg
property instead of size and including the entire available area into
ro-boot. Note that since CBFS definitions are not being changed, the
actual use of the RO section is not changing.

The upper half includes two 1.5MB areas, each accommodating a key
block, a rw section and some additional areas as required. 1MB on top
remains reserved for legacy firmware.

BRANCH=storm
BUG=chrome-os-partner:34161
TEST=built and booted a full storm image.

Change-Id: Iac75764efc32f886dad1e66dd0f1033dce535d0a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236961
Reviewed-by: David Hendricks <dhendrix@chromium.org>