nand: Insert dummy partitions for NAND numbering gap
In practice, partition 6 and 7 are often blank. MTD partitioning numbers
sequentially enumerating over all the partitions that are created. If blank
partitions are left out, it would make a numbering mismatch between the kernel
and the GPT intentions--the OEM partition would look like mtd6. There was code
in depthcharge which kills the boot when it detects this mismatch. This patch
replaces that behavior with inserting empty partitions in the gap, so that
there are dummy partitions mtd6 and mtd7 of size 0 and partition 8 is sized
and numbered appropriately.

BUG=chromium:444726
TEST=booted with partitions 1-5 and 8 created and observed all partitions to
come up in /proc/mtd with the proper numbering, including 6 and 7 as empty.
root@localhost $ cat /proc/mtd
dev:    size   erasesize  name
mtd0: 40000000 00040000 "device"
mtd1: 26840000 00040000 "STATE"
mtd2: 01000000 00040000 "KERN-A"
mtd3: 08600000 00040000 "ROOT-A"
mtd4: 01000000 00040000 "KERN-B"
mtd5: 08600000 00040000 "ROOT-B"
mtd6: 00000000 00000000 "blank"
mtd7: 00000000 00000000 "blank"
mtd8: 00c00000 00040000 "OEM"
BRANCH=none

Change-Id: I332eb9e8a94879391aba19a7ab6303038adf2900
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237307
Reviewed-by: Julius Werner <jwerner@chromium.org>