flash: Implement erase on SPI NOR flash
This patch implements 'erase' on SPI NOR flash devices. Erase-related
parameters, namely the erase command and erase granularity ('sector
size'), come from coreboot/libpayload. While we're at it, get the size
of the flash device from there also, so it doesn't have to be passed
in when initializing NAND. The Storm board had been using only part
of the flash in order to save memory; this patch just increases the
heap size to make it fit.

BUG=chromium:446377
BRANCH=none
TEST=With the following patch which implements SPI GPT rewriting,
verified that the 'tries' count is decremented on boot. This was
run on a Storm Proto 0.2 board, which includes a Spansion part
(S25FL256S) with a 64k sector size.
TEST=Verified that the daisy board still compiles with the change
CQ-DEPEND=CL:239570

Change-Id: If637e1086805e1933f89733fa21991f0104c5e4c
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238733
Reviewed-by: Julius Werner <jwerner@chromium.org>