nand: Read-modify-write GPT in SPI NOR flash
This patch implements a read-modify-erase-write scheme for external
calls to VbExDiskWrite to SPI GPT devices. Previously, erases were
not done so writes just resulted in data corruption.

During boot, vboot modifies the GPT for several reasons:
- The tries count is decremented so that if boot leads to a
  crash, there is only a finite number of retries until switching
  to the other kernel or recovery.
- If verification fails, then as an optimization, priority and
  other fields are set to 0 to avoid reverification.
Therefore, the firmware needs to write to NOR. It would be possible
to design another system where these are records in nvstorage, but
I believe sticking with GPT is simpler and remains low-overhead.

To make room for the larger GPT, the main RW boot firmware is shrunken
by 64k. It seems to still fit in that space.

BUG=chromium:446377
BRANCH=none
TEST=Observed repeated boots to decrement the 'tries' count and set
an unverified kernel to all 0. Repeated boots did not corrupt the
GPT.

Change-Id: I18148f64c586783736b4c3e3e608cfc2226522a5
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238734