sound: Add codec driver for rt5677
NOTE: This is added directly to the samus firmware branch so it
does not conflict with ongoing work to enable rt5677 on TOT for
other platforms.

I started with the same driver that Nvidia is working on for Ryu
and then made a few changes:

1) remove debug code
2) remove some unused code (samus just uses MCLK instead of the PLL)
3) a few 80 column line cleanups
4) a few necessary register settings to work with Intel SST

The changes I made to work with Intel SST will need to be integrated
with the TOT driver that is being developed for Ryu but there is a
more urgent need for this codec diver on samus so we can write protect
the firmware soon.

BUG=chrome-os-partner:29650
BRANCH=samus
TEST=boot on samus in developer mode, check for beep after 30 seconds

Change-Id: I6e5cc88b7466b0ab9a96fd43f8fa9cdb116e8311
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238432
Reviewed-by: Tom Warren <twarren@nvidia.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>