storm: Split up GPT into sections for piecewise rewriting
When userspace rewrites the GPT, it should do so separately for the
first half and second half of the GPT, erasing and rewriting them
one at a time so that, if it is interrupted by an unclean shutdown,
one copy of the GPT is still valid. This patch splits up the RW_GPT
section into RW_GPT_PRIMARY and RW_GPT_SECONDARY which can be
addressed individually. It maintains the combined RW_GPT section as
it was which everyone else can use as before.

TEST=builds
BUG=chromium:442518
BRANCH=none

Change-Id: I5300add4573d9eeee904eed024bc4dfed564d0d6
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238881
Reviewed-by: Julius Werner <jwerner@chromium.org>