storm: Double depthcharge heap size
Storm includes a 16MiB flash device (at least on the proto 0.2
board), but its heap is allocated in depthcharge to be 16MiB big.
To work around this, since depthcharge allocates memory the size
of the whole flash device up front, the flash device is initialized
as an 8 MiB big device. An upcoming change to the way flash is
used in depthcharge will get the actual size from coreboot, from
probing the device, so this technique will no longer work.

This patch increases the size of the heap to 32MiB in order to
hold the real size of the flash plus other things that need to
be allocated during boot. The device has more than enough memory
to accommodate this allocation.

Maybe we should allocate memory to back flash from elsewhere;
this is tracked by chromium:447758.

BRANCH=none
BUG=chromium:446377
BUG=chromium:447758
TEST=booted with the following series of patches on proto 0.2 and
saw the memory allocated for flash, and then the kernel loaded
correctly from both NAND and USB storage

Change-Id: Id1b9ee032b38c1ddc092ae0ec65b3a110fac8710
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/239940
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>