intel i915: Add display driver to turn off panel
To prevent possible garbage between the firmware screens and the
chrome os splash screen have depthcharge turn off the panel if
it is on and let the kernel power it back up.

BUG=chrome-os-partner:28234
BRANCH=broadwell
TEST=build and boot on samus in dev mode and check that the panel
does not show garbage before the chrome os splash screen.  Also
ensure this code does not run in normal or legacy boot paths.

Change-Id: I60f46c5eb4ce5afbb0720c0d25e00fc512c06f77
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240842
Reviewed-by: Shawn N <shawnn@chromium.org>