nand: Disable SPI GPT in recovery mode
Recovery mode must not read data in read-write until it gets onto the
verified USB stick, as that could lead to a bricked device case. This
includes the GPT held on SPI NOR flash. This patch disables the GPT
on SPI, which would otherwise happen when reading the partition table
into the device tree.

BUG=chromium:448620
TEST=Hacked src/vboot/stages.c to simulate a recovery mode and found
that the GPT on SPI is not read when booting from either USB or internal
storage with the recovery mode simulation, but normal GPT reading is
still working without that.
BRANCH=none

Change-Id: I829b01ceebc10a53941dd12bc8c353e60a0d3dbb
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243460
Reviewed-on: https://chromium-review.googlesource.com/261702
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>