ryu: Increase heap size to fix flash allocation issue
A recent change in how flash is initialized made the flash have the
actual size reported by the device, rather than a size hard-coded
into board.c. Ryu had falsely reported a smaller flash size in board.c
and its heap that was too small relative to the actual flash to
allocate memory to back it. This patch bumps up the Ryu heap size
so that the allocation of memory for flash doesn't fail.

BRANCH=none
BUG=chromium:446377
TEST=Not yet well-tested

Change-Id: Ie82e119afa49227f036928ff7efefdb5310436c0
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241433
Reviewed-by: Aaron Durbin <adurbin@chromium.org>