tegra: handle all DMA constraints
The DMA engine on tegra requires the pointers be 4 byte aligned
and the size of the transfers be in multiples of 4 bytes. Update
the code to assert on these constraints as well as handle those
constraints by ensuring DMA transfers are greater than or equal
to the cacheline size. This assumes the cacheline size is larger
than the 4 byte DMA constraints as well as handles the situation
of proceeding and succeeding memory of the buffer is used in the
course of a transfer because of the buffer alignment. The remaining
possibilty is a prefetcher pulling cache lines back into cache
prior to read being completed.

BUG=None
BRANCH=None
TEST=Built and can boot ryu again.

Change-Id: I60da0eca73416ec6c3dee100a0fcb54c35709ebb
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241515
Reviewed-by: Julius Werner <jwerner@chromium.org>