bdev_ctrlr: Add ops to identify if a block is owned by a ctrlr
Currently, there is no way to establish if a particular block dev is
owned by a particular controller. In order to establish the relation
between them, add a function is_bdev_supported to bdev_ctrlr ops. Bdev
ctrlr checks to see if it owns the block device and returns 1 on
success and 0 on failure.

BUG=chrome-os-partner:35511
BRANCH=None
TEST=Compiles successfully. Able to verify using fastboot protocol
that the correct block device is identified for a ctrlr.

Change-Id: I54144bdc64bfd15edef81a1159e186d886ce9866
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/241730
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>