intel_i915: Turn off backlight instead of sequencing panel down
Some panels (particularly with PSR) are not retraining cleanly
in the kernel if we sequence it down in firmware.  Since I cannot
be sure what the VBIOS is doing with PSR just turn off the backlight.

BUG=chrome-os-partner:35352
BRANCH=broadwell
TEST=Boot on multiple systems and make sure the screen does not
show garbage after dev screen and before kernel splash screen.
Also ensure that PSR panels do not have any issues when they come
up in the OS.

Change-Id: Ie98c203644cf2295aed570f33aa8c911d63ab252
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241765
Reviewed-by: Shawn N <shawnn@chromium.org>