depthcharge: add support for architecture specific ldscript
The MIPS architecture needs specific symbols and regions added to the
linker script in order for depthchearge to be built and to execute
properly. Currently depthcharge uses an architecture independent
linker script, that cannot be used for this purpose.
Therefore, in this implementation, LDSCRIPT is set to use either
the architecture specific ldscript found in src/arch/$(ARCH_DIR)
or the default one from depthcharge, based on a Kconfig variable.

BUG=chrome-os-partner:31438
TEST=tested on Pistachio FPGA and bring up board; depthcharge is built
     and executed properly, when using an architecture specific
     linker script.
BRANCH=none

Change-Id: I49ba0f7b4436c9f9cd6823dc4d338b8cc66bcac2
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/242391
Reviewed-by: David Hendricks <dhendrix@chromium.org>