arch/mips: add MIPS specific linker script
The MIPS architecture needs specific symbols and regions added to the
linker script in order for depthchearge to be built and to execute
properly. Currently depthcharge uses an architecture independent
linker script, that cannot be used for this purpose.
Additional to the default linker script, this one defines a _gp symbol,
necessary for gp-relative addressing and an exception section with the
code in exception_asm.S which is best placed at BASE_ADDRESS as it
needs a 4KB alignment.
Also, scommon section was added to bss in order for small common
symbols to be cleared in startup code.

BUG=chrome-os-partner:31438
TEST=tested on Pistachio FPGA and bring up board; depthcharge is built
     and executed properly, when using this linker script.
BRANCH=none

Change-Id: I0248dc0de8930b51b10a090340125e035bc8dfc0
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/243273
Reviewed-by: David Hendricks <dhendrix@chromium.org>