mips: Pass flattened device-tree to kernel
The kernel expects the physical address of a flattened device-tree
blob to be passed in by the bootloader, but depthcharge passes in
an unflattened device-tree.  Fix this by passing the FDT blob instead.

Note that, at the moment, virt_to_phys() does not actually convert
the address to a physical one, so the kernel will still die early
in the boot process because it expected the physical FDT address.

BUG=chrome-os-partner:31438
BRANCH=none
TEST=With a hack to not convert the firmware-provided FDT address
to a virtual address, the kernel is able to scan the FDT.

Change-Id: I6edbd330790a2eb2bd4040eac3d92579d5a43dce
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/245536
Reviewed-by: David Hendricks <dhendrix@chromium.org>