depthcharge: spi-nand: Rework SPI NAND command transfer
Using a separate transfers for the command and the address might
result in random I/O errors. This commit reworks the way SPI messages
are prepared for each command, putting the command and the address
in a single transfer.

While here, add some sanity checks on the command.

BUG=chrome-os-partner:31438
TEST=tested on Pistachio bring up board; works as expected;
BRANCH=none

Author: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Change-Id: Ibbaf9783913fc38325dcf5c25b0bebdaf01b951d
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/246481
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>