nand: Treat -EUCLEAN as success for stream reads
When mtd has a read which is over the ECC threshold, the result
is -EUCLEAN, meaning the data was read properly but it is a warning
to consider moving the data. When booting, this message should be
ignored. This patch changes the treatment of -EUCLEAN from fatal
error to ignored in mtd streaming reads.

Also a minor cleanup: \n at the end of some print statements.

BUG=chrome-os-partner:31438
TEST=With bad block skipping turned off, managed to read a bit
from SPI NAND, which has a very EUCLEAN-happy driver, reporting
that error code when any bits are corrected.
BRANCH=none

Change-Id: I9e70ce55373e108e2229a927e0a2ba3d6d4cbfd4
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/246463
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>