spi-nand: Fix address calculation on read
In the SPI NAND protocol, a 24-bit address is used for reads from NAND
to the read cache, with units in terms of pages (writesize). The driver
was shifting addresses as if they were in units of bytes. This patch
fixes the SPI NAND driver to use page size for reads.

BUG=chrome-os-partner:35348
TEST=Together with some other workarounds (disabling bad block skipping
and SPI timeouts), depthcharge successfully read, verified and jumped
to the kernel on NAND
BRANCH=none

Change-Id: I8d6e43d0808cc09a8162ce552485656f5429c133
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/247830
Reviewed-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Tested-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>