Disable boot retry countdown if lid is closed
This avoids problems when the closed system is repeatedly
powered on, where the OS might shut down before registering
the successful boot.

Instead, just boot opportunistically in that case. If things
don't work, the counter will take care of that when booting
with the lid open.

CQ-DEPEND=CL:249150
BRANCH=none
BUG=chromium:446945
TEST=Start closed device by giving it external power and wait
for power-down repeatedly. Check that it doesn't fall back.

Change-Id: I245054b8e7efd1db531ac49542e2e7f1fc6f1dd9
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/250800
Reviewed-by: Randall Spangler <rspangler@chromium.org>