ipq806x: i2c: maintain a list of available interfaces
With upcoming designs there is a need to keep track of multiple i2c
interfaces.

Depthcharge already provides a means of abstract lists of items. This
patch adds the ability to add i2c interfaces to a list as well.
Presently this is meant to be used only when CLI is enabled.

So far it has been the board file responsibility to include the
interfaces into the lists, which causes a lot of code duplication.

This patch makes this the responsibility of the controller driver, as
soon as i2c interface is initialized, its description is added to the
interface list. Currently only the ipq806x i2c driver adds its
instances to the list, other drivers could follow suite as required.

The driver generates the interface name to be used as the interface
handle, this could be changed: for instance have the caller
initializing the interface pass in the desired name.

The actual function adding interfaces to the list will be submitted in
the following patch, as it is required only when CLI is enabled.

BRANCH=storm
BUG=chrome-os-partner:36057
TEST=verified that the interfaces are added to the list as expected
     and are available to the CLI

Change-Id: I7961ead53c64150a3191358e5ea31126e6ed09e0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/249442
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/252163