fastboot ryu: Provide mode entry indication
In order to enable developers to understand that fastboot mode is
entered on the device, enable display and print a message on the
screen. This is just a temporary fix until fastboot flows are
finalized and we decide how to let the user interact with fastboot
modes.

BUG=chrome-os-partner:35861
BRANCH=None
TEST=Prints "Entered fastboot mode" message on scren

Change-Id: I5981816a2a159c7214f8933c0c33116684acec30
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/249791
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>