i2c: IMGTEC driver implementation
This is a driver for the IMG serial control bus (SCB) IP block.
The driver is a simplified version of the Linux driver for this
block, having the I2C controller driven in automatic mode, with
state control guided by polling the interrupt registers.

In automatic mode, a bus address, a data count and whether to
read or write is specified, and the hardware takes care of the
I2C state machine, using a FIFO to send/receive bytes of data
to/from an I2C slave. The driver has to keep the FIFO drained
or filled in response to the values of the interrupt registers.

BRANCH=none
BUG=chrome-os-partner:31438
TEST=tested on Pistachio bring up board; works properly with the
     TPM
     - when CLI is compiled in and the driver is enabled, it shows up
       in the 'i2c list' output, but fails to scan the bus on
       Concerto.
     - when enabled on concerto board, the driver still fails to
       communicate with the TPM

Change-Id: I3a1a49226496ead119a0a9a2b5670cea8aa85795
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/250490