mips: vboot: use the same find_common_params() on all but X86 platforms
There are two versions of the find_common_params() for non-unified
depthcharge, one for x86 platforms, ACPI based, and another one - for
ARM platforms, which uses a global variable. Now there is a need to
support this on MIPS platforms. Instead of duplicating ARM
implementation in the MIPS tree, generalize the ARM implementation to
be used on all but X86 platforms.

BRANCH=none
BUG=chrome-os-partner:31438
TEST=compiled depthcharge for samus and storm

Change-Id: I7173e5170b10d3ca2b854d9df14486b8bff99f69
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/250322
Reviewed-by: Aaron Durbin <adurbin@chromium.org>