storm: enable dev mode keyboard support
It has been long contended that being able to control storm device
behavior in dev mode through keyboard would be beneficial.

It turns out that enabling USB keyboard also has a desirable side
effect of priming the USB stick, so that they are ready to go in case
the user wants to boot from USB.

To support the headless device behavior let's just keep polling the
physical presence button when polling the keyboard in dev mode. The
"not pressed->pressed" transition of the phys presence button is
equivalent of pressing CTL^U on the keyboard.

BRANCH=storm
BUG=chrome-os-partner:36059
TEST=with the USB input configuration option enabled the USB stick
     initializes soon after boot and pressing of the dev mode button
     triggers a successful attempt of booting kernel from USB.

Change-Id: I60450db318ddaf431276f82a3398ece94ca7e89f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/252424
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Conflicts:
	src/vboot/callbacks/headless_stub.c
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/325830