ipq806x: power: implement power callbacks
Depthcharge needs to be able to reboot the device on the vboot layer
request. This patch introduces power management callbacks for the
ipq806x SOC.

The reset function is the same as in coreboot, the power off function
is not yet implemented (it is not even clear if it is be necessary)
and currently invokes the reset function.

BRANCH=storm
BUG=chrome-os-partner:36059

TEST=with this config enabled tried booting with corrupted kernel,
     observed the target reboot in recovery mode.

Change-Id: Ia2d62a4ec65a129bc4d6e8b206ac4378e9d7629a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/253090
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/254596