use kseg0 to wipe out low memory ranges
The lowest 1MByte of DRAM on MIPS platforms is protected from direct
access, but this memory needs to be wiped out by the firmware. Luckily
MIPS architecture provides multiple apertures to access same memory.
Modify the memory set routines such that in case the region starts
within the first 1MByte of address range, the KSEG0 window is used.

This patch assumes that there will be no wipeout requests wrapping
around the memory address range and that the address range will not
exceed the KSEG0 size, both of these conditions are guaranteed to be
true at current state of the codebase.

BRANCH=none
BUG=chrome-os-partner:36258
TEST=ran on urara, verified that the wipeout happens successfully and
     the board boots all the way to the Linux login prompt.

Change-Id: Iff6cbac9392b2849341c18f47f683edb33968679
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/254314
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>