i2c: allow to suppress write error messages
When an i2c bus is scanned, the code attempts to read register zero at
all bus device addresses. Each of these read attempts starts with an
i2c write transaction when the register address is written into the
device.

In case of accessing non-existing devices the i2c controller observes
a NACK and reports it on the console. This floods the console with
error messages, as usually very few devices are attached to any i2c
bus.

This patch introduces a callback and a state variable which would
allow the clients using the i2c interface to suppress error messages
generated during write accesses.

BRANCH=storm
BUG=chrome-os-partner:37223
TEST=with this callback supported in the ipq806x i2c driver, the
     following is reported during the gsbi1 bus scan on sp5 (the
     reported errors are still being investigated):
 storm: i2c pick gsbi1
 Will use bus gsbi1
 storm: i2c scan
 Will scan bus gsbi1
 qup_i2c_write() returns 11
 qup_i2c_write() returns 11
 qup_i2c_write() returns 11
 qup_i2c_write() returns 11
 0x20
 storm:

Change-Id: I558099b86720e289756176f864cbabd983d1ddb1
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/254910
Reviewed-by: Aaron Durbin <adurbin@chromium.org>