lp55231: Implement LED ring driver
This is a driver for the Whirlwind LED ring, which is equipped with
two LED microcontrollers Texas Instruments LP55231
(http://www.ti.com/product/lp55231), each of them driving three
multicolor LEDs.

The only connection between the ring and the main board is an i2c bus.

Even though technically this is an LED device, it is capable of
displaying quite sophisticated patterns and is supposed to display
distinctly different patterns at different states of verified boot. so
this driver imitates a depthcharge display device. On initialization
the driver sets up the controllers to prepare them to accept programs
to run.

When a certain vboot state needs to be indicated, the program for that
state is loaded into the controllers, resulting in the state
appropriate LED behavior.

The way to reset a controller is to write 0xff into register 0x3d. The
controller resets while the last byte of this i2c transaction is still
in transit, and this results in a NAK during last byte's transmission.

The ipq806x i2c driver fails to report this error, but reports an
error on the following i2c bus access attempt. To work around this the
driver first writes the register, and then reads it, ignoring errors
on both transactions.

The ww_ring_display_screen() driver method tries to find a program
matching a vboot state and if successful - loads the program into the
controllers, resulting in the desired LED behavior.

The actual programs for vboot states are not yet available, the sample
program included in the driver blinks two LEDs on the ring.

BRANCH=storm
BUG=chrome-os-partner:37223, chrome-os-partner:36059
TEST=with all patches applied, the SP5 board on boot initializes the
     ring to blink two red LEDs.

Change-Id: Icecc28f4f6b34068f283851a81c5c1d21757b6ec
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256403
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261785