urara: determine hardware properties at run time
The concerto board (a urara derivative) can potentially be used in
many different hardware configurations. The same firmware can also be
used on Urara bring up board (BuB). The following properties could be
different for different configurations:

 - SPI amd NAND chip select GPIOs
 - i2c interfaces
 - kernel FDT compatible strings

The board ID retrieved by coreboot and communicated through sysinfo
table is a unique descriptor of the board and the devices it is
running, this value can be used to pick the correct set of properties
when initializing firmware.

The included mapping table covers currently existing concerto
derivatives and the Urara BuB.

BRANCH=none
BUG=chrome-os-partner:37593
TEST=manual
  - verified that changing board ID reported by coreboot causes change
    in the compatible string and SPI NAND chip select.

Change-Id: I07ae171d19ffee55adca8512e57874d9d94f2e20
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/257181
Reviewed-by: David Hendricks <dhendrix@chromium.org>