net : ipq806x : fix ethernet for Whirlwind
Modify the ipq806x ethernet switch driver not to hardcode the
interface to initialize, but to retrieve the port number from the
board file.

Also adapt the driver to depthcharge driver model (functions included
through INIT_FUNC() are not supposed to touch the hardware).

Use the poll callback for actual driver initialization.

BUG=chrome-os-partner:37442
BRANCH=master
TEST=netboot on Whirlwind and storm
     . Fetched vmlinuz and cmdline from tftp server.

Change-Id: I26672bd36e98be90e5a896b0ebfc545f8e291669
Signed-off-by: Yogesh Lal <ylal@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/259488
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>