net: do not assume there is a single usb based network interface
Depthcharge presumes that there could be only one ehternet device
present in the system, and that device is a USB dongle. This
assumption does not hold true for some newer devices in development.

This patch decouples usb stack from the network driver and introduces
two new lists, one to allow to register drivers which need polling,
and one to allow to register Net interfaces (as opposed to allowing a
single one).

The network driver provides just one function to let the app wait for
network availability, net_wait_for_link(). This function keeps polling
and keeps checking status of the registered interfaces. As soon as an
interface reports link up, the function makes that interface the
default networking interface and exits.

BRANCH=storm
BUG=chrome-os-partner:28915
TEST=tested on Whirlwind with the built in interface disabled. The
     driver initializes and attempts TFTP boot:

  Waiting for link
  Initializing XHCI USB controller at 0x11000000.
  Adding net device
  done.
  MAC: c0:c1:c0:76:38:7e
  Sending DHCP discover... done.
  Waiting for reply... done.
  Sending DHCP request... done.
  Waiting for reply... done.
  My ip is 192.168.1.63
  The DHCP server ip is 192.168.1.1
  TFTP server IP supplied by DHCP server: 192.168.1.1
  Bootfile supplied by DHCP server:

Change-Id: I66a4bda8684f3d92ba9f0d2e1fa2fa3bbc40e79d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/260030