net: ipq806x: use SOC resources to determine MAC address
When the device comes from the production line its VPD area is not yet
configured, which means there is no valid MAC address in the sysinfo
table when the Ethernet driver is initialized.

Luckily, the ipq906x SOC provides two registers which allow to
retrieve some unique for the SOC 64 bit number.

The default MAC address is a combination of 0x02 in the first byte
(which means that this is a locally administered MAC address) and the
top 5 bytes of the sha1 calculated over the 64 but number retrieved
from the SOC.

BRANCH=storm
BUG=chrome-os-partner:37444, chrome-os-partner:37444
TEST=verified that on three different VPD-less storm/whirlwind devices
     the MAC address is set to unique consistent value, maintained
     over resets.

Change-Id: Id37ff1b48532e4c81056284c18726a5344f1328a
Signed-off-by: Sourabh Banerjee <sbanerje@codeaurora.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/259583
Reviewed-by: Hung-Te Lin <hungte@chromium.org>