cli: i2c: usability improvements
The i2c CLI command expects the user to pick an interface to operate
on. But on many systems there is only one i2c interface, so the
explicit pick is not required. Change the behavior to use the first
available interface is there has been no one picked yet.

Also, allow the 'write' option to pass multiple data parameters.

BRANCH=none
BUG=none
TEST=verified multibyte write capability loading programs into the WW
     LED ring controllers.

Change-Id: I198d67a2a12e7ebd1758f3a3036e09b3875ee391
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/260497
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261793