ww_ring: define display pattern programs
Add compiled lp55231 code snippets to allow display certain patterns
corresponding to various vboot screens.

The behavior of the WW ring is determined by the jetstream PRD.

The patch also includes the source code for the LED controller as well
as instructions on how to compile and modify the code to result in
different colors, intensities, blink periods and duty cycles.

BRANCH=storm
BUG=chrome-os-partner:36059
TEST=with the added CLI command observed different patterns for
     different vboot conditions

Change-Id: I72e698fdaaeb9c45795b7643ce48f9b61a4719be
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/260707
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261795