audio: provide audio stack volume control ability
Some audio implementations have volume control capabilities. It is
convenient to be able to change volume at run time when trying to
debug audio features. Let's add plumbing to allow this if the driver
supports it.

The new API allows to set the volume level in abstract units in 0..100
range, the driver is expected to normalize this value.

BRANCH=none
BUG=chrome-os-partner:38048
TEST=with volume control plugged in in ipq806x, it is possible to
     change the volume of the speaker beep.

Change-Id: I5e41efef26e706518f16b8d503a57317300c61f9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261037
Reviewed-by: Aaron Durbin <adurbin@chromium.org>