rockchip: Remove unnecessary settings from spi_start()
The Rockchip SPI driver currently reinitializes CTRLR0, even though it
should have already been set up by coreboot. This is unnecessary and
dangerous because settings might go out of sync. spi_start() is also
invoked on every transfer (not just once), and writing this register
over and over again is excessive.

This patch removes those lines so that spi_start() will only perform the
one purpose it's supposed to, asserting CS. It also removes the
appropriate arguments from new_rockchip_spi() which concern parameters
that should've already been configured by coreboot. This even includes
the CS parameter, since the code currently doesn't even use it right...
(it ignores it and always writes a '1' to the SER register). We can add
it back in if we ever want to fix this, but for now it's better to leave
it out than to have it non-functional.

BRANCH=veyron
BUG=chrome-os-partner:38352
TEST=Booted Jerry.

Change-Id: I4cee7f81f6bff8339c33a36ade4e3c6d667f9b50
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262311
Reviewed-by: David Hendricks <dhendrix@chromium.org>