dt: function to copy WiFi calibration data into the device tree
This patch generalizes the function previously used by storm for
copying WiFi calibration data from the sysinfo table into the device
tree.

Until now storm allowed for both binary and base64 calibration data
storage format in the VPD and the device tree. Binary format is
preferable, but the VPD does not provide a means of storing binary
data.

With the recent addition of the base64 decoder in the coreboot, the
calibration data in sysinfo is stored in binary, and this obviates the
need for the device tree entries to carry base64 formatted data.

On both storm and urara WiFi calibration data blobs are now placed
into the respective device tree locations in binary.

BRANCH=storm
BUG=chromium:450169
TEST=verified that on both storm and urara the appropriate device tree
     nodes are created and the binary calibration data matches the
     base64 VPD contents.

Change-Id: Iddee1fc881720376aefcdc7773cd01a31a6093e3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/263021
Reviewed-by: Aaron Durbin <adurbin@chromium.org>