vboot: provide ability to allow USB boot on normal to dev transitions
Storm devices do not allow the user to access the device when they
transition from normal to developer mode. At the very least the user
should be allowed to boot their own system image from the USB stick.

The way to achieve this is to set the NVRAM flag allowing to boot from
USB in developer mode. The new configuration option, when enabled,
will signal the vboot layer that the NVRAM flag should be set.

CQ-DEPEND=CL:264187
BRANCH=storm
BUG=chrome-os-partner:38303
TEST=tested with the next patch, which enables this feature on storm

Change-Id: Ib3f843f62e890d9f29019e8152fab7bb1589b20f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/264213
Reviewed-by: Bill Richardson <wfrichar@chromium.org>